# Optimize for ASCII-only strings without control symbols