@janlelis 2023-01-02
@fatkodima Great suggestion and thanks for testing.

The suggestion solution is a great starting point, and maybe we can go with it. But one thing I ask myself, would it increase the time needed for strings w/ around 10000 chars that do contain Unicode characters notably?

To answer your question: It's absolutely OK to use this newer method. I'll just add a note to the README that people stuck on an older Ruby can just use an older version of this gem.

I will also experiment a little over the next days and maybe merge/release around Friday