@fatkodima 2023-01-02
Benchmarked for `str = "a" * 10_000 + "\n"`:
```
Warming up --------------------------------------
            original    10.000  i/100ms
           optimized    10.000  i/100ms
Calculating -------------------------------------
            original    100.018  (± 7.0%) i/s -    500.000  in   5.029530s
           optimized     99.432  (± 5.0%) i/s -    500.000  in   5.046291s

Comparison:
            original:      100.0 i/s
           optimized:       99.4 i/s - same-ish: difference falls within error
```