@janlelis 2023-01-03
Hi @fatkodima,

the new version with your performance optimization is released! I have only tweaked the regex a little to cherry-pick specific control characters, instead of prohibiting all of them. 

Btw, I would have loved to use a single regex match (`/[[[:^ascii:]]\0\x05\a\b\n\v\f\r\x0E\x0F]/`) instead of using the `#ascii_only?` method, but that did not play well together with UTF16 strings... Would have had a similar performance, so it doesn't matter anyways... :thinking: